+++ app/src/main/res/layout/word_list.xml
@@ -5,5 +5,5 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.android.miwok.NumbersActivity"
+    tools:context="com.example.android.miwok.MainActivity"
     android:drawSelectorOnTop="true"/>
